Refactors Cached files and the TabBar
