+++ frontend/src/js/components/modals/CachedFiles/Tabs/TabBar.jsx
@@ -1,16 +1,22 @@
 import React, { PureComponent } from 'react'
 import { Button } from 'react-bootstrap'
 
+const tabs = {
+    file: "file",
+    esgf: "esgf",
+    opendap: "opendap",
+    info: "info"
+}
+
 class TabBar extends PureComponent{
 
     render(){
         return(
             <div className="text-center variable-tabs">
-                <Button id="tabbar-file" bsSize="small" bsStyle={this.props.selectedTab == "file" ? 'primary' : 'default'} onClick={() => {this.props.switchTab('file')}}>File</Button>
-                <Button id="tabbar-esgf" disabled bsSize="small" bsStyle={this.props.selectedTab == "esfg" ? 'primary' : 'default'} onClick={() => {this.props.switchTab('esgf')}}>ESGF</Button>
-                <Button id="tabbar-opendap" disabled bsSize="small" bsStyle={this.props.selectedTab == "opendap" ? 'primary' : 'default'} onClick={() => {this.props.switchTab('opendap')}}>OpenDAP</Button>
-                <Button id="tabbar-edit" disabled bsSize="small" bsStyle={this.props.selectedTab == "edit" ? 'primary' : 'default'} onClick={() => {this.props.switchTab('edit')}}>Edit</Button>
-                <Button id="tabbar-info" bsSize="small" bsStyle={this.props.selectedTab == "info" ? 'primary' : 'default'} onClick={() => {this.props.switchTab('info')}}>Info</Button>
+                <Button id="tabbar-file" bsSize="small" bsStyle={this.props.selectedTab == tabs.file ? 'primary' : 'default'} onClick={() => {this.props.switchTab('file')}}>File</Button>
+                <Button id="tabbar-esgf" disabled bsSize="small" bsStyle={this.props.selectedTab == tabs.esfg ? 'primary' : 'default'} onClick={() => {this.props.switchTab('esgf')}}>ESGF</Button>
+                <Button id="tabbar-opendap" disabled bsSize="small" bsStyle={this.props.selectedTab == tabs.opendap ? 'primary' : 'default'} onClick={() => {this.props.switchTab('opendap')}}>OpenDAP</Button>
+                <Button id="tabbar-info" bsSize="small" bsStyle={this.props.selectedTab == tabs.info ? 'primary' : 'default'} onClick={() => {this.props.switchTab('info')}}>Info</Button>
             </div>
         )
     }
@@ -20,4 +26,6 @@ TabBar.propTypes = {
     switchTab: React.PropTypes.func,
     selectedTab: React.PropTypes.string,
 }
+
 export default TabBar
+export { tabs }
